Implement authorization using Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,7 @@ use App\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
 
+
 class QuestionsController extends Controller
 {
     /**
@@ -66,7 +67,12 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question',$question))
+        {
+            abort(403, "Access denied!");
+        }
         return view("questions.edit", compact('question'));
+        
     }
 
     /**
@@ -78,6 +84,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question',$question))
+        {
+            abort(403, "Access denied!");
+        }
+        
         $question->update($request->only('title','body'));
 
         return redirect()->route('questions.index')->with('success',"Your question has been updated");
@@ -91,6 +102,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question))
+        {
+            abort(403, "Access denied!");
+        }
+
         $question->delete();
 
        return redirect()->route('questions.index')->with('success',"Your question has been deleted.");
